+++ packages/framework/src/Model/Order/Item/OrderProductService.php
@@ -7,18 +7,6 @@ class OrderProductService
     /**
      * @param \Shopsys\FrameworkBundle\Model\Order\Item\OrderProduct[] $orderProducts
      */
-    public function subtractOrderProductsFromStock(array $orderProducts)
-    {
-        $orderProductsUsingStock = $this->getOrderProductsUsingStockFromOrderProducts($orderProducts);
-        foreach ($orderProductsUsingStock as $orderProductUsingStock) {
-            $product = $orderProductUsingStock->getProduct();
-            $product->subtractStockQuantity($orderProductUsingStock->getQuantity());
-        }
-    }
-
-    /**
-     * @param \Shopsys\FrameworkBundle\Model\Order\Item\OrderProduct[] $orderProducts
-     */
     public function returnOrderProductsToStock(array $orderProducts)
     {
         $orderProductsUsingStock = $this->getOrderProductsUsingStockFromOrderProducts($orderProducts);
@@ -47,7 +35,7 @@ class OrderProductService
      * @param \Shopsys\FrameworkBundle\Model\Order\Item\OrderProduct[] $orderProducts
      * @return \Shopsys\FrameworkBundle\Model\Order\Item\OrderProduct[]
      */
-    private function getOrderProductsUsingStockFromOrderProducts(array $orderProducts)
+    public function getOrderProductsUsingStockFromOrderProducts(array $orderProducts)
     {
         $orderProductsUsingStock = [];
         foreach ($orderProducts as $orderProduct) {
