+++ packages/framework/tests/Unit/Model/Order/Item/OrderProductServiceTest.php
@@ -12,48 +12,6 @@ use Shopsys\FrameworkBundle\Model\Product\ProductData;
 
 class OrderProductServiceTest extends TestCase
 {
-    public function testSubtractOrderProductsFromStockUsingStock()
-    {
-        $productStockQuantity = 15;
-        $orderProductQuantity = 10;
-
-        $orderMock = $this->createMock(Order::class);
-
-        $productData = new ProductData();
-        $productData->usingStock = true;
-        $productData->stockQuantity = $productStockQuantity;
-        $product = Product::create($productData);
-        $productPrice = new Price(0, 0);
-
-        $orderProduct = new OrderProduct($orderMock, 'productName', $productPrice, 0, $orderProductQuantity, null, null, $product);
-
-        $orderProductService = new OrderProductService();
-        $orderProductService->subtractOrderProductsFromStock([$orderProduct]);
-
-        $this->assertSame($productStockQuantity - $orderProductQuantity, $product->getStockQuantity());
-    }
-
-    public function testSubtractOrderProductsFromStockNotUsingStock()
-    {
-        $productStockQuantity = 15;
-        $orderProductQuantity = 10;
-
-        $orderMock = $this->createMock(Order::class);
-
-        $productData = new ProductData();
-        $productData->usingStock = false;
-        $productData->stockQuantity = $productStockQuantity;
-        $product = Product::create($productData);
-        $productPrice = new Price(0, 0);
-
-        $orderProduct = new OrderProduct($orderMock, 'productName', $productPrice, 0, $orderProductQuantity, null, null, $product);
-
-        $orderProductService = new OrderProductService();
-        $orderProductService->subtractOrderProductsFromStock([$orderProduct]);
-
-        $this->assertSame($productStockQuantity, $product->getStockQuantity());
-    }
-
     public function testAddOrderProductsToStockUsingStock()
     {
         $productStockQuantity = 15;
